API

Constructing models

VCSEL.VCModel — Type

VCModel(Y, X, V, Σ)
VCModel(Y, X, V)
VCModel(Y, V, Σ)
VCModel(Y, V)

Variance component model type. Stores the data and model parameters of a variance component model.

VCSEL.VCintModel — Type

VCintModel(Y, X, V, Vint, Σ, Σint)
VCintModel(Y, X, V, Vint)
VCintModel(Y, V, Vint, Σ, Σint)
VCintModel(Y, V, Vint)

Variance component interaction model type. Stores the data and model parameters of a variance component model.

Properties of model

Base.length — Method

length(vcm::VCModel)
length(vcm::VCintModel)

Length d of response.

VCSEL.ncovariates — Function

ncovariates(vcm::VCModel)
ncovariates(vcm::VCintModel)

Number of fixed effects parameters p of a VCModel or VCintModel.

Base.size — Method

size(vcm::VCModel)
size(vcm::VCintModel)

Size (n, d) of response matrix of a VCModel or VCintModel.

VCSEL.nmeanparams — Method

nmeanparams(vcm::VCModel)
nmeanparams(vcm::VCintModel)

Number of mean parameters p * d of VCModel or VCintModel.

VCSEL.nvarcomps — Method

nvarcomps(vcm::VCModel)
nvarcomps(vcm::VCintModel)

Number of variance components.

VCSEL.ngroups — Method

ngroups(vcm::VCModel)
ngroups(vcm::VCintModel)

Number of groups, m, for VCModel or VCintModel.

Fitting model

VCSEL.vcselect! — Function

vcselect!(vcm; penfun, λ, penwt, maxiters, tol, verbose)

Input

vcm: VCModel.

Keyword Argument

penfun: penalty function. Default is NoPenalty().

λ: tuning parameter. Default is 1.

penwt: penalty weights. Default is (1,...1,0).

standardize: logical flag for covariance matrix standardization. Default is false. If true, V[i] is standardized by its Frobenius norm, and parameter estimates are returned on the original scale.

maxiters: maximum number of iterations. Default is 1000.

tol: convergence tolerance. Default is 1e-6.

verbose: display switch. Default is false.

checktype: check argument type switch. Default is true.

Output

vcm: VCModel with updated Σ and β.

obj: objective value at convergence.

niters: number of iterations to convergence.

objvec: vector of objvective values at each iteration.

Example

vcm1 = VCModel(Y, X, V)
vcselect!(vcm1; penfun=L1Penalty(), λ=1.5)

vcselect!(vcm; penfun, λ, penwt, maxiters, tol, verbose)

Input

vcm: VCintModel

Keyword

penfun: penalty function (e.g. NoPenalty(), L1Penalty(), MCPPenalty()), default is NoPenalty()

λ: tuning parameter, default is 1.0

penwt: weights for penalty term, default is (1,1,...1,0)

maxiters: maximum number of iterations, default is 1000

standardize: logical flag for covariance matrix standardization, default is false. If true, V[i] and Vint[i] is standardized by its Frobenius norm, and parameter estimates are returned on the original scale

tol: convergence tolerance, default is 1e-5

verbose: display switch, default is false

checktype: check argument type switch, default is true

objvec: vector of objvective values at each iteration

Output

vcm: VCintModel with updated Σ, Σint and β Access estimates with vcm.Σ, vcm.Σint and vcm.β

obj: objective value at convergence

niters: number of iterations to convergence

Example

vcm = VCModel(Y, X, V, Vint)
vcselect!(vcm; penfun=MCPPenalty(), λ=1.5)

VCSEL.vcselectpath! — Function

vcselectpath!(vcm; penfun=NoPenalty(), penwt=[ones(nvarcomps(vcm)-1); 0], nλ=100, 
                λpath=zeros(0), maxiters=1000, tol=1e-6, verbose=false)

Generate solution path of variance components along varying lambda values.

Input

vcm: VCModel.

Keyword

penfun: penalty function (e.g. NoPenalty(), L1Penalty(), MCPPenalty()). Default is NoPenalty().

penwt: weights for penalty term. Default is (1,1,...1,0).

nλ: number of tuning parameter values. Default is 100.

λpath: a user supplied λ sequence in ascending order. Typically the program computes its own λ sequence based on nλ; supplying λpath overrides this.

maxiter: maximum number of iteration for MM loop. Default is 1000.

standardize: logical flag for covariance matrix standardization. Default is false. If true, V[i] is standardized by its Frobenius norm.

tol: convergence tolerance. Default is 1e-6.

Output

Σ̂path: matrix of estimated variance components at each tuning parameter λ. Each column gives vector of estimated variance components σ2 at certain λ.

β̂path: matrix of fixed parameter estimates at each λ.

λpath: sequence of λ values used.

objpath: vector of objective values at each tuning parameter λ.

niterspath: vector of no. of iterations to convergence.

Example

vcm = VCModel(Y, X, V)
Σ̂path, β̂path, λpath, objpath, niterspath = vcselectpath!(vcm; penfun=MCPPenalty(), nλ=50)

vcselectpath!(vcm; penfun, penwt, nλ, λpath, maxiters, tol, verbose)

Generate solution path of variance components along varying lambda values.

Input

vcm: VCintModel

Keyword

penfun: penalty function (e.g. NoPenalty(), L1Penalty(), MCPPenalty()), default is NoPenalty()

penwt: weights for penalty term, default is (1,1,...1,0)

nλ: number of tuning parameter values, default is 100

λpath: a user supplied λ sequence. Typically the program computes its own λ sequence based on nλ; supplying λpath overrides this

maxiter: maximum number of iteration for MM loop, default is 1000

standardize: logical flag for covariance matrix standardization, default is false. If true, V[i] and Vint[i] are standardized by its Frobenius norm

tol: convergence tolerance, default is 1e-5

verbose: display switch, default is false

Output

Σ̂path: matrix of estimated variance components at each tuning parameter λ, each column gives vector of estimated variance components σ2 at certain λ

Σ̂intpath: matrix of estimated variance components at each tuning parameter λ, each column gives vector of estimated variance components σ2 at certain λ

β̂path: matrix of fixed parameter estimates at each λ; i-th column is an estimate at λpath[i]

λpath: sequence of λ values used

objpath: vector of objective values at each λ; i-th value corresponds to λpath[i]

niterspath: vector of no. of iterations to convergence at each λ; i-th value corresponds to λpath[i]

Example

vcm = VCModel(Y, X, V, Vint)
Σ̂path, Σ̂intpath, = vcselectpath!(vcm; penfun=L1Penalty(), λpath=range(0, 10.0, length=50))

Utilities

VCSEL.resetModel! — Function

resetModel!(vcm::VCModel)

Reset VCModel with initial estimates Σ. If Σ is unspecified, it is set to a vector of ones or identity matrices based on its dimension.

resetModel!(vcm::VCModel, Σ)

Reset VCModel with initial estimates Σ.

resetModel!(vcm::VCintModel, [Σ, Σint])

Reset VCintModel with initial estimates Σ and Σint (if given). If unspecified, it is set to a vector of ones.

VCSEL.maxlambda — Function

maxlambda(y, V; penfun=NoPenalty(), penwt=[ones(length(V)-1); 0.0], maxiters=500, 
          standardize=false, tol=1e-8)

Find the value of λ where all σ turns 0. At any value greater than λ, all σ's are 0.

Input

y: response vector

V: vector of covariance matrices

Keyword

penfun: penalty function, possible options are L1Penalty() (default), MCPPenalty(), and NoPenalty(). If penfun=NoPenalty(), 0 is returned

penwt: vector of penalty weights, default is (1,1,...1,0)

standardize: logical flag for covariance matrix standardization, default is false. If true, V[i] and Vint[i] is standardized by its Frobenius norm, and parameter estimates are returned on the original scale

maxiters: maximum number of iterations, default is 500

tol: value below which estimate is considered zero

Output

maxλ: value of λ where all coeffecients turn zero, maximum λ on the solution path

maxlambda(Y, V; penfun=NoPenalty(), penwt=[ones(length(V)-1); 0.0], standardize=false,
          maxiters=500, tol=1e-8)

Find the value of λ where all σ turns 0. At any value greater than λ, all σ's are 0.

Input

Y: response matrix

V: vector of covariance matrices

Keyword

penfun: penalty function, possible options are L1Penalty() (default), MCPPenalty(), and NoPenalty(). If penfun=NoPenalty(), 0 is returned

penwt: vector of penalty weights, default is (1,1,...1,0)

standardize: logical flag for covariance matrix standardization, default is false. If true, V[i] and Vint[i] is standardized by its Frobenius norm, and parameter estimates are returned on the original scale

maxiters: maximum number of iterations, default is 500

tol: value below which estimate is considered zero

Output

maxλ: value of λ where all coeffecients turn zero, maximum λ on the solution path

maxlambda(y, V, Vint; penfun=NoPenalty(), penwt=[ones(length(V)-1); 0.0], 
          maxiters=500, tol=1e-8)

Find the value of λ where all σ turns 0. At any value greater than λ, all σ's are 0.

Input

y: response vector

V: vector of covariance matrices

Vint: vector of covariance matrices for interaction

Keyword

penfun: penalty function, possible options are L1Penalty() (default), MCPPenalty(), and NoPenalty(). If penfun=NoPenalty(), 0 is returned

penwt: vector of penalty weights, default is (1,1,...1,0)

standardize: logical flag for covariance matrix standardization, default is false. If true, V[i] and Vint[i] is standardized by its Frobenius norm, and parameter estimates are returned on the original scale

maxiters: maximum number of iterations, default is 500

tol: value below which estimate is considered zero

Output

maxλ: value of λ where all coeffecients turn zero, maximum λ on the solution path

VCSEL.rankvarcomps — Function

rankvarcomps(Σpath; tol=1e-8)

Input

Σpath: solution path (in numeric matrix), each column should represent estimated variance components at specific λ as in output from vcselect!, vcselectpath!

resvarcomp: logical flag indicating the presence of residual variance component in Σpath. If true, last row of Σpath is ignored. Default is true

Keyword Argument

tol: value such that variance components less than this value is considered zero, default is 1e-8

Output

ranking: rank of each variance component based on the order in which it enters solution path

rest: rest of the variance components that are estimated to be zero at all λ > 0

Example

vcm = VCModel(Y, X, V)
Σ̂path, = vcselectpath!(vcm; penfun=MCPPenalty(), nλ=50)
posvarcomps, = rankvarcomps(Σ̂path)

rankvarcomps(Σpath; tol=1e-8)

Input

Σpath: solution path (matrix of matrices), each column should contain estimated variance component matrices at specific λ as in output from vcselectpath!

resvarcomp: logical flag indicating the presence of residual variance component in Σpath. If true, last row of Σpath is ignored. Default is true

Keyword Argument

tol: value such that a variance component matrix whose trace is less than this value is considered zero, default is 1e-8

Output

ranking: rank of each variance component based on the order in which it enters solution path

rest: rest of variance components that are estimated to be zero at all λ > 0

rankvarcomps(Σpath, Σpath2; tol=1e-6, resvarcomp=true, resvarcomp2=true)

Obtain rank of variance components from a paired solution paths, e.g. Σpath and Σpath2. Ranks are calculated using norm of paired variance components.

Input

Σpath: solution path (in numeric matrix), each column should represent estimated variance components at specific λ as in output from vcselect, vcselectpath

Σpath2: solution path (in numeric matrix), each column should represent estimated variance components at specific λ as in output from vcselect, vcselectpath

Keyword

tol: a variance component less than tol is considered zero, default is 1e-6

resvarcomp: logical flag indicating there is residual variance component in Σpath, default is true. If true, the last variance component is not included in ranks

resvarcomp2: logical flag indicating there is residual variance component in Σpath2, default is true. If true, the last variance component is not included in ranks

Output

ranks: rank of each variance component based on the order in which it enters solution path

rest: rest of the variance components that are estimated to be zero at all λ > 0

Example

vcm = VCModel(Y, X, V, Vint)
Σ̂path, Σ̂intpath, = vcselectpath!(vcm; penfun=L1Penalty(), λpath=range(0, 10.0, length=50))
posvarcomps, zerovarcomps = rankvarcomps(Σ̂path, Σ̂intpath; tol=1e-5)

VCSEL.plotsolpath — Function

plotsolpath(Σpath, λpath; title="Solution Path", xlab="λ", ylab="σ2", 
        xmin=minimum(λpath), xmax=minimum(λpath), tol=1e-6)

Output plot of solution path at varying λ values. Use backend such as gr().

Input

Σpath: solution path (in numeric matrix) to be plotted, each column should represent variance components at specific λ as in output from vcselectpath!. The last row of Σpath should be for residual variance component. Otherwise set resvarcomp=false to indicate the absence.

λpath: vector of tuning parameter λ values.

Keyword

title: title of the figure. Default is "Solution Path".

xlab: x-axis label. Default is "lambda".

ylab: y-axis label. Default is "sigma_i^2".

xmin: lower limit for x-axis. default is minimum of λpath.

xmax: upper limit for x-axis. default is maximum of λpath.

linewidth: line width. Default is 1.0.

nranks: no. of ranks to displayed on legend. Default is total number of variance components.

legend: logical flag for including legend. Default is true.

legendout: logical flag for moving the legend outside the plot. Default is true.

legendtitle: legend title. Default is "Ranking".

resvarcomp: logical flag for indicating residual variance component in Σpath. Default is true.

Output

plot of solution path.

plotsolpath(Σpath, λpath; title="Solution Path", xlab="λ", ylab="σ2", 
        xmin=minimum(λpath), xmax=minimum(λpath), tol=1e-6)

Output plot of solution path at varying λ values. Use backend such as gr().

Input

Σpath: solution path (in numeric matrix) to be plotted, each column should represent variance components at specific λ as in output from vcselectpath!. The last row of Σpath should be for residual variance component. Otherwise set resvarcomp=false to indicate the absence.

λpath: vector of tuning parameter λ values.

Keyword

title: title of the figure. Default is "Solution Path".

xlab: x-axis label. Default is "lambda".

ylab: y-axis label. Default is "sigma_i^2".

xmin: lower limit for x-axis. default is minimum of λpath.

xmax: upper limit for x-axis. default is maximum of λpath.

linewidth: line width. Default is 1.0.

nranks: no. of ranks to displayed on legend. Default is total number of variance components.

legend: logical flag for including legend. Default is true.

legendout: logical flag for moving the legend outside the plot. Default is true.

legendtitle: legend title. Default is "Ranking".

resvarcomp: logical flag for indicating residual variance component in Σpath. Default is true.

Output

plot of solution path.

plotsolpath(Σpath, Σpath2, λpath; title="Solution Path", xlab="λ", ylab="σ2", 
        xmin=minimum(λpath), xmax=minimum(λpath), tol=1e-6)

Output plot of a paired solution path at varying λ values. Use backend such as gr().

Input

Σpath: solution path (in numeric matrix) to be plotted, each column should represent variance components at specific λ as in output from vcselect, vcselectpath

Σpath2: solution path (in numeric matrix) to be plotted, each column should represent variance components at specific λ as in output from vcselect, vcselectpath

λpath: vector of tuning parameter λ values

Keyword

title: title of the figure. Default is "Solution Path".

xlab: x-axis label. Default is "lambda".

ylab: y-axis label. Default is "sigma_i^2".

xmin: lower limit for x-axis. default is minimum of λpath.

xmax: upper limit for x-axis. default is maximum of λpath.

linewidth: line width. Default is 1.0.

nranks: no. of ranks to displayed on legend. Default is total number of variance components.

legend: logical flag for including legend. Default is true.

legendout: logical flag for moving the legend outside the plot. Default is true.

legendtitle: legend title. Default is "Ranking".

resvarcomp: logical flag for indicating residual variance component in Σpath. Default is true.

Output

plot of solution path